Account - READ COMMENT
Right. For the account to work right now you have to manually input the
price you bought shares for, and the amount you made when you sell
shares, manually. This is because we aren't getting the messages from
the server through the interface, they're just being printed out in
receivemessages. If you want the automation you'll have to change
everything so we can get the message in interface and to get the price
and profit automatically.

+++ src/StructNode.java
@@ -3,10 +3,10 @@ public class StructNode {
 	
 	public String company;
 	public int ownedShares;
-	public float sharePrice;
+	public double sharePrice;
 	public StructNode next;
 
-	public StructNode(String companyName, int numberOfShares, float price)
+	public StructNode(String companyName, int numberOfShares, double price)
 	{
 		company = companyName;
 		ownedShares = numberOfShares;
